Better collision detection Add enemies of no enemies are left

+++ intersection.py
@@ -1,7 +1,7 @@
 import shapely.geometry
 
 
-def doesIntersect(shotStart: tuple, shotEnd: tuple, boxStart: tuple, width: int, height: int) -> bool:
+def doesIntersect(shotStart: tuple, shotEnd: tuple, boxStart: tuple, width: float, height: float) -> bool:
     poly = shapely.geometry.Polygon([boxStart,
                                      (boxStart[0] + width, boxStart[1]),
                                      (boxStart[0] + width, boxStart[1] + height),
